Add files via upload
This file includes three folders that are named on the basis of geophysical models. Each folder (model) has images derived from the inversion process that can be used as initial data for programming. In addition to an Excel file, there are two Python files that were written to produce the final results.